Wired Forecast Activity and HomeActivity

+++ data/src/main/java/com/example/data/dto/ForecastResponse.kt
@@ -45,7 +45,7 @@ data class Main(
     val pressure: Double,
     val sea_level: Double,
     val temp: Double,
-    val temp_kf: Int,
+    val temp_kf: Double,
     val temp_max: Double,
     val temp_min: Double
 )
